🔧 Fix image extensions to use enum
