+++ packages/myst-cli/src/process/site.ts
@@ -20,7 +20,8 @@ import { filterPages, loadProjectFromDisk } from '../project';
 import { castSession } from '../session';
 import type { ISession } from '../session/types';
 import { watch, selectors } from '../store';
-import { ImageExtensions, transformWebp } from '../transforms';
+import { transformWebp } from '../transforms';
+import { ImageExtensions } from '../utils';
 import { combineProjectCitationRenderers } from './citations';
 import { loadIntersphinx } from './intersphinx';
 import type { PageReferenceStates, TransformFn } from './mdast';
